Refactored the code to revert to the original csv format and data, and use the user id to present as the user name
Added routes from HomeActivity to the QuestionnaireActivity and InsightsActivity that are appropriate to the app.

Refactored the intent flags for various routes to work accordingly for the app

Implemented the base UI code for the QuestionnaireActivity, to have the various buttons and components for the user to fill in. No backend code put in place yet